stock_broker/pyproject.toml
[project]
name = "stock_broker"
version = "0.1.0"
description = "stock-broker using crewAI"
authors = [{ name = "Ashish shukla", email = "ashishshukla1810@gmail.com" }]
requires-python = ">=3.10,<3.14"
dependencies = [
    "crewai[tools]>=0.157.0,<1.0.0",
    "pandas>=1.5.0",
    "numpy>=1.24.0", 
    "requests>=2.28.0",
    "python-dotenv>=0.19.0"
]

[project.scripts]
stock_broker = "stock_broker.main:run"
run_crew = "stock_broker.main:run"
train = "stock_broker.main:train"
replay = "stock_broker.main:replay"
test = "stock_broker.main:test"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.crewai]
type = "crew"


